Calculate the branching instruction offset relative to the branched target instruction offset.
This is also reported in issue GH-1 on GitHub.

Previous implementation: Assign the absolute instruction offset as the
branching instruction offset when a label is referenced.

New implementation: Set the branching instruction offset relative to the
branched target instruction offset when a label is referenced.

The eBPF JIT compiler is supposed to skip the instruction at the specified
program counter offset in a branching instruction. Providing `0` as the offset
will cause the flow of execution to resume by executing the next instruction.

Consider the following examples:

```
PC:  |  Instructions:  |
        loop0:
1.      add r2 1
2.      sub r1 1
3.      jge r1 r2 loop0
```

The assembler will replace `loop0` with `-3` for the label reference in the
above code snippet.

```
PC:  |  Instructions:  |
        loop1:
1.      jlt r1 r2 loop2
2.      add r2 1
3.      sub r1 1
        loop2:
4.      jge r2 r1 loop1
```

The assembler will replace `loop2` with `2` and `loop1` with `-4`, for the
label references in the above code snippet.

--------

The BPF ISA specifies that:
The offset to increment the program counter by is in units of 64-bit
instructions relative to the instruction following the jump instruction.
* `PC += 1`  skips execution of the next instruction.
* `PC += -1` re-executes the current instruction.
where altering the program counter results in undefined behavior if the next
instruction is a 128-bit wide instruction.

Signed-off-by: Emil Masoumi <emilmasoumi@protonmail.com>